pousuwane zbędne komentarze i importy. Typy posegregowane
