+++ src/types/ListTypes.ts
@@ -18,39 +18,7 @@ export type ListItem = {
     completionTime: string | null
 }
 
-export type RegistrationData = {
-    email: string;
-    name: string;
-    surname: string;
-    phone?: string | null,
-    password: string;
-    terms: boolean;
-}
 
-export type LoginInputs = {
-    email?: string,
-    password: string,
-}
-export type LogourData = {
-    _id: string,
-}
-
-export type UserType = {
-    email: string
-    name: string
-    surname: string,
-    _id: string
-    phone: string
-    role: UserRoles
-}
-export type EditUserType = {
-    _id?: string,
-    name: string,
-    surname: string,
-    email: string,
-    password?: string,
-    phone?: string | null
-}
 
 export type TaskItem = {
     userId: string;
@@ -59,12 +27,6 @@ export type TaskItem = {
     description: string;
     category: TaskCategory,
     status: 'completed' | 'in-progress' | "active";
-}
-
-export enum UserRoles {
-    ADMIN = 'admin',
-    MANAGER = 'manager',
-    EMPLOYEE = 'employee'
 }
 
 // powinnismy miec rozne pliki dla roznych typow typow
